+++ src/Command/CacheCommand.php
@@ -39,7 +39,7 @@ class CacheCommand extends AbstractLockedCommand
     protected function executeLocked(InputInterface $input, OutputInterface $output)
     {
         $this->io      = new SymfonyStyle($input, $output);
-        $this->rootDir = dirname($this->getContainer()->getParameter('kernel.root_dir'));
+        $this->rootDir = $this->getContainer()->getParameter('kernel.project_dir');
 
         $files = [
             '/system/tmp/stylesheet-manager',